utils audit
